pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "A modern FastAPI template with clean architecture"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.127.0,<0.128.0",
    "uvicorn>=0.40.0,<0.41.0",
    "sqlalchemy>=2.0.45,<3.0.0",
    "pydantic>=2.12.5,<3.0.0",
    "pydantic-settings>=2.12.0,<3.0.0",
    "asyncpg>=0.31.0,<0.32.0",
    "alembic>=1.17.2,<2.0.0",
    "httpx>=0.28.1,<0.29.0",
    "fastapi-pagination>=0.15.4,<0.16.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "passlib[bcrypt]>=1.7.4,<2.0.0",
    "python-jose[cryptography]>=3.3.0,<4.0.0",
    "python-multipart>=0.0.9,<0.1.0"
]

[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = "A modern FastAPI template with clean architecture"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.127.0"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.45"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
asyncpg = "^0.31.0"
alembic = "^1.17.2"
httpx = "^0.28.1"
fastapi-pagination = "^0.15.4"
python-dotenv = "^1.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
black = "^24.0.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
